Authentication

+++ routes/reviewRoute.js
@@ -5,6 +5,7 @@ const ExpressError = require("../utils/ExpressError");
 const {reviewSchemaJoi} = require("../schema.js");
 const Review = require("../models/reviewSchema");
 const Story = require("../models/storySchema");
+const { getAuth } = require("@clerk/express");
 
 // Joi Validation
 const validateReview = (req, res, next) => {
@@ -17,12 +18,14 @@ const validateReview = (req, res, next) => {
 }
 
 // Reviews Post route
-router.post("/",validateReview, wrapAsync(async(req, res) => {
-  let story = await Story.findById(req.params.id);
-  let newReview = new Review(req.body.review);
-
+router.post("/", validateReview, wrapAsync(async(req, res) => {
+  const { userId } = getAuth(req);
+  const story = await Story.findById(req.params.id);
+  const newReview = new Review({
+    ...req.body.review,
+    author: userId,
+  });
   story.reviews.push(newReview);
-
   await newReview.save();
   await story.save();
   res.redirect(`/stories/${story._id}`);
@@ -30,11 +33,17 @@ router.post("/",validateReview, wrapAsync(async(req, res) => {
 
 // Reviews Delete route
 router.delete("/:reviewId", wrapAsync(async(req, res) => {
-  let {id, reviewId} = req.params;
+  const { userId } = getAuth(req);
+  const {id, reviewId} = req.params;
+  const review = await Review.findById(reviewId);
+  
+  if(!review || review.author !== userId){
+    return res.redirect(`/stories/${id}`);
+  }
   await Story.findByIdAndUpdate(id, {$pull: {reviews: reviewId}});
   await Review.findByIdAndDelete(reviewId);
-
   res.redirect(`/stories/${id}`);
 }));
+
 
 module.exports = router;
