+++ routes/storyRoute.js
@@ -4,6 +4,13 @@ const wrapAsync = require("../utils/wrapAsync");
 const ExpressError = require("../utils/ExpressError");
 const {storySchemaJoi} = require("../schema.js");
 const Story = require("../models/storySchema");
+const { clerkClient, getAuth } = require("@clerk/express");
+
+function isLoggedIn(req, res, next) {
+  const { userId } = getAuth(req);
+  if (!userId) return res.redirect("/signin");
+  next();
+}
 
 // Joi Validation
 const validateStory = (req, res, next) => {
@@ -16,16 +23,21 @@ const validateStory = (req, res, next) => {
 }
 
 // Create route GET
-router.get("/create", (req, res) => {
+router.get("/create", isLoggedIn, (req, res) => {
     res.render("stories/createStory.ejs");
 });
 
 // Create route POST
-router.post("/create", validateStory, wrapAsync(async (req, res) => {
-  const newStory = new Story(req.body.story);
+router.post("/create", isLoggedIn, validateStory, wrapAsync(async (req, res) => {
+  const { userId } = getAuth(req);
+  const newStory = new Story({
+    ...req.body.story,
+    author: userId,
+  });
   await newStory.save();
   res.redirect("/stories");
 }));
+
 
 // All stories
 router.get("/", async (req, res) => {
@@ -36,30 +48,50 @@ router.get("/", async (req, res) => {
 // Show route
 router.get("/:id", wrapAsync(async (req, res, next) => {
   const { id } = req.params;
+  const { userId } = getAuth(req);
   const story = await Story.findById(id).populate("reviews");
-  res.render("stories/showStory", { story });
+   for (let review of story.reviews) {
+    if (review.author) {
+      review.user = await clerkClient.users.getUser(review.author);
+    }
+  }
+  res.render("stories/showStory", { story, userId });
 }));
 
 // Edit route
-router.get("/:id/edit", wrapAsync(async (req, res) => {
-  let { id } = req.params;
+router.get("/:id/edit", isLoggedIn, wrapAsync(async (req, res) => {
+  const { id } = req.params;
   const story = await Story.findById(id);
+  if(story.author !== getAuth(req).userId) {
+    return res.redirect(`/stories/${id}`);
+  }
   res.render("stories/edit.ejs", { story });
 }));
 
+
 // Update route
-router.put("/:id", validateStory, wrapAsync(async (req, res) => {
-  const { id } = req.params;
-  const { title, category, story } = req.body;
-  await Story.findByIdAndUpdate(id, { title, category, story });
+router.put("/:id", isLoggedIn, validateStory, wrapAsync(async (req,res)=>{
+  const {id} = req.params;
+  const story = await Story.findById(id);
+  if(story.author !== getAuth(req).userId) {
+    return res.redirect(`/stories/${id}`);
+  }
+  await Story.findByIdAndUpdate(id, req.body.story);
   res.redirect(`/stories/${id}`);
 }));
 
+
+
 // Delete route
-router.delete("/:id", wrapAsync(async (req, res) => {
-  const { id } = req.params;
+router.delete("/:id", isLoggedIn, wrapAsync(async (req,res)=>{
+  const {id} = req.params;
+  const story = await Story.findById(id);
+  if(story.author !== getAuth(req).userId) {
+    return res.redirect(`/stories/${id}`);
+  }
   await Story.findByIdAndDelete(id);
   res.redirect("/stories");
 }));
+
 
 module.exports = router;
